requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-gis==0.16
maps==5.1.1
pytz==2020.4
sqlparse==0.4.1
